editing ingredients

+++ src/components/Ingredients/Ingredients.js
@@ -2,6 +2,7 @@ import React, { useReducer, useEffect, useCallback, useMemo } from 'react';
 
 import IngredientForm from './IngredientForm';
 import IngredientList from './IngredientList';
+import EditModal from '../UI/EditModal';
 import ErrorModal from '../UI/ErrorModal';
 import Search from './Search';
 
@@ -15,6 +16,14 @@ const ingredientReducer = (ingredientState, action) => {
       return [...ingredientState, action.ingredient];
     case 'DELETE':  // Remove an ingredient from the current state
       return ingredientState.filter(ingredient => ingredient.id !== action.id)
+    case 'EDIT':
+      return ingredientState.map((ingredient => {
+        if (ingredient.id === action.id) {
+          ingredient.title = action.data.title;
+          ingredient.amount = action.data.amount;
+        }
+        return ingredient;
+      }));
     default:
       throw new Error(`Missing type in ingredientReducer: ${action.type}`);
   }
@@ -30,7 +39,7 @@ const initialEditState = {
 const editReducer = (editState, action) => {
   switch (action.type) {
     case 'EDIT':
-      return {inUse: true, ...action.editFields}
+      return { inUse: true, ...action.editFields }
     case 'RESET':
       return initialEditState
   }
@@ -39,18 +48,25 @@ const editReducer = (editState, action) => {
 const Ingredients = () => {
   const [ingredientsState, ingredientsDispatch] = useReducer(ingredientReducer, []);
   const [editState, editDispatch] = useReducer(editReducer, initialEditState);
-  const { data, error, id, ingredient, loading, resetError, sendRequest } = useHttp();
+  const { data, edit, error, id, ingredient, loading, resetError, sendRequest } = useHttp();
 
   useEffect(() => {
-    if (id && !error) {
+    if (id && !edit && !error) {
       ingredientsDispatch({ type: 'DELETE', id: id });
     } else if (ingredient && data && !error) {
       ingredientsDispatch({
         type: 'ADD',
         ingredient: { id: data.name, ...ingredient }
       });
+    } else if (data && id && edit && !error) {
+      ingredientsDispatch({
+        type: 'EDIT',
+        id: id,
+        data: data
+      });
+      resetEditHandler();
     }
-  }, [data, id, ingredient]);
+  }, [data, edit, id, ingredient]);
 
   // useCallback allows React to cache the function so it survives re-render cycles. 
   // This causes the function not to change and prevents infinite loops.
@@ -66,7 +82,7 @@ const Ingredients = () => {
   }, []);
 
   const editHandler = useCallback((id, title, amount) => {
-    editDispatch({ 
+    editDispatch({
       type: 'EDIT',
       editFields: {
         id: id,
@@ -76,12 +92,16 @@ const Ingredients = () => {
     });
   }, []);
 
+  const editIngredientHandler = useCallback((ingredient, id) => {
+    sendRequest(`https://react-hooks-1d6ad.firebaseio.com/ingredients/${id}.json`, 'PATCH', JSON.stringify(ingredient), id, null, true);
+  }, [sendRequest]);
+
   const addIngredientHandler = useCallback((ingredient) => {
-    sendRequest('https://react-hooks-1d6ad.firebaseio.com/ingredients.json', 'POST', JSON.stringify(ingredient), null, ingredient);
+    sendRequest('https://react-hooks-1d6ad.firebaseio.com/ingredients.json', 'POST', JSON.stringify(ingredient), null, ingredient, null);
   }, [sendRequest]);
 
   const removeIngredientHandler = useCallback((id) => {
-    sendRequest(`https://react-hooks-1d6ad.firebaseio.com/ingredients/${id}.json`, 'DELETE', null, id, null);
+    sendRequest(`https://react-hooks-1d6ad.firebaseio.com/ingredients/${id}.json`, 'DELETE', null, id, null, null);
   }, [sendRequest]);
 
   const resetErrorMessageHandler = useCallback(() => {
@@ -105,6 +125,8 @@ const Ingredients = () => {
   return (
     <div className="App">
       {error && <ErrorModal onClose={resetErrorMessageHandler}>{error}</ErrorModal>}
+      {editState.inUse && <EditModal onClose={resetEditHandler} editFields={editState} onEditIngredient={editIngredientHandler}></EditModal>}
+
       <IngredientForm
         onAddIngredient={addIngredientHandler}
         loading={loading}
@@ -113,12 +135,12 @@ const Ingredients = () => {
       <section>
         <Search onLoadIngredients={filteredIngredientsHandler} />
         <IngredientList
-        ingredients={ingredientsState}
-        onRemoveItem={removeIngredientHandler}
-        onEdit={editHandler}
-        resetEdit={resetEditHandler}
-        editState={editState}
-      />
+          ingredients={ingredientsState}
+          onRemoveItem={removeIngredientHandler}
+          onEdit={editHandler}
+          resetEdit={resetEditHandler}
+          editState={editState}
+        />
       </section>
     </div>
   );
